LPS-114383 Move GenericUtil to petra-reflect
