+++ modules/apps/content-dashboard/content-dashboard-web/build.gradle
@@ -10,5 +10,6 @@ dependencies {
 	compileOnly project(":apps:frontend-taglib:frontend-taglib-clay")
 	compileOnly project(":apps:frontend-taglib:frontend-taglib-react")
 	compileOnly project(":core:osgi-service-tracker-collections")
+	compileOnly project(":core:petra:petra-reflect")
 	compileOnly project(":core:registry-api")
 }
