+++ modules/apps/info/info-impl/build.gradle
@@ -6,4 +6,5 @@ dependencies {
 	compileOnly project(":apps:info:info-api")
 	compileOnly project(":apps:static:osgi:osgi-util")
 	compileOnly project(":core:osgi-service-tracker-collections")
+	compileOnly project(":core:petra:petra-reflect")
 }
