+++ modules/apps/info/info-impl/src/main/java/com/liferay/info/internal/display/contributor/InfoDisplayContributorTrackerImpl.java
@@ -16,12 +16,12 @@ package com.liferay.info.internal.display.contributor;
 
 import com.liferay.info.display.contributor.InfoDisplayContributor;
 import com.liferay.info.display.contributor.InfoDisplayContributorTracker;
-import com.liferay.info.internal.util.GenericsUtil;
 import com.liferay.info.item.provider.InfoItemFormProvider;
 import com.liferay.info.item.provider.InfoItemProvider;
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
 import com.liferay.osgi.util.ServiceTrackerFactory;
+import com.liferay.petra.reflect.GenericsUtil;
 
 import java.util.ArrayList;
 import java.util.Dictionary;
