+++ modules/apps/info/info-impl/src/main/java/com/liferay/info/internal/item/selector/InfoItemSelectorTrackerImpl.java
@@ -14,9 +14,9 @@
 
 package com.liferay.info.internal.item.selector;
 
-import com.liferay.info.internal.util.GenericsUtil;
 import com.liferay.info.item.selector.InfoItemSelector;
 import com.liferay.info.item.selector.InfoItemSelectorTracker;
+import com.liferay.petra.reflect.GenericsUtil;
 import com.liferay.portal.kernel.util.Validator;
 
 import java.util.ArrayList;
