+++ modules/apps/info/info-impl/src/main/java/com/liferay/info/internal/util/ItemClassNameServiceReferenceMapper.java
@@ -16,6 +16,7 @@ package com.liferay.info.internal.util;
 
 import com.liferay.osgi.service.tracker.collections.map.PropertyServiceReferenceMapper;
 import com.liferay.osgi.service.tracker.collections.map.ServiceReferenceMapper;
+import com.liferay.petra.reflect.GenericsUtil;
 
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.ServiceReference;
