+++ modules/core/petra/petra-reflect/src/main/resources/com/liferay/petra/reflect/packageinfo
@@ -1,1 +1,1 @@
-version 1.3.0
+version 1.4.0
