+++ modules/dxp/apps/analytics/analytics-reports-impl/build.gradle
@@ -3,5 +3,6 @@ dependencies {
 	compileOnly group: "org.osgi", name: "org.osgi.annotation.versioning", version: "1.1.0"
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
+	compileOnly project(":core:petra:petra-reflect")
 	compileOnly project(":dxp:apps:analytics:analytics-reports-api")
 }
